+++ package.json
@@ -19,7 +19,7 @@
     "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
     "mongodb": "^3.1.6",
-    "mongoose": "^7.4.2",
+    "mongoose": "^5.2.17",
     "mysql2": "^1.6.1",
     "pug": "^2.0.3",
     "sequelize": "^5.0.0-beta.11"
