--- public/controllers/error.js
@@ -1,3 +1,0 @@
-exports.get404 = (req, res, next) => {
-  res.status(404).render('404', { pageTitle: 'Page Not Found', path: '/404' });
-};
