--- public/data/cart.json
@@ -1,1 +1,0 @@
-{"products":[{"id":"0.41607315815753076","qty":1}],"totalPrice":12}
