--- public/public/css/cart.css
@@ -1,25 +1,0 @@
-.cart__item-list {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  margin: auto;
-  width: 40rem;
-  max-width: 90%;
-}
-
-.cart__item {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.26);
-  margin-bottom: 1rem;
-}
-
-.cart__item h1,
-.cart__item h2 {
-  margin-right: 1rem;
-  font-size: 1.2rem;
-  margin: 0;
-}
-
