--- public/public/css/forms.css
@@ -1,23 +1,0 @@
-.form-control {
-  margin: 1rem 0;
-}
-
-.form-control label,
-.form-control input,
-.form-control textarea {
-  display: block;
-  width: 100%;
-  margin-bottom: 0.25rem;
-}
-
-.form-control input,
-.form-control textarea {
-  border: 1px solid #a1a1a1;
-  font: inherit;
-  border-radius: 2px;
-}
-
-.form-control input:focus,
-.form-control textarea:focus {
-  outline-color: #00695c;
-}
