--- public/public/css/product.css
@@ -1,27 +1,0 @@
-.product-form {
-    width: 20rem;
-    max-width: 90%;
-    margin: auto;
-    display: block;
-}
-
-.product-item {
-    width: 20rem;
-    max-width: 95%;
-    margin: 1rem;
-}
-
-.product__title {
-    font-size: 1.2rem;
-    text-align: center;
-}
-
-.product__price {
-    text-align: center;
-    color: #4d4d4d;
-    margin-bottom: 0.5rem;
-}
-
-.product__description {
-    text-align: center;
-}
