--- public/util/path.js
@@ -1,3 +1,0 @@
-const path = require('path');
-
-module.exports = path.dirname(process.mainModule.filename);
