--- public/views/404.ejs
@@ -1,8 +1,0 @@
-<%- include('includes/head.ejs') %>
-</head>
-
-<body>
-    <%- include('includes/navigation.ejs') %>
-    <h1>Page Not Found!</h1>
-
-<%- include('includes/end.ejs') %>
