--- public/views/includes/add-to-cart.ejs
@@ -1,4 +1,0 @@
-<form action="/cart" method="post">
-    <button class="btn" type="submit">Add to Cart</button>
-    <input type="hidden" name="productId" value="<%= product._id %>">
-</form>
