--- public/views/includes/end.ejs
@@ -1,4 +1,0 @@
-<script src="/js/main.js"></script>
-</body>
-
-</html>
