+++ routes/admin.js
@@ -7,18 +7,18 @@ const adminController = require('../controllers/admin');
 const router = express.Router();
 
 // /admin/add-product => GET
-// router.get('/add-product', adminController.getAddProduct);
+router.get('/add-product', adminController.getAddProduct);
 
-// // /admin/products => GET
-// router.get('/products', adminController.getProducts);
+// /admin/products => GET
+router.get('/products', adminController.getProducts);
 
-// // /admin/add-product => POST
-// router.post('/add-product', adminController.postAddProduct);
+// /admin/add-product => POST
+router.post('/add-product', adminController.postAddProduct);
 
-// router.get('/edit-product/:productId', adminController.getEditProduct);
+router.get('/edit-product/:productId', adminController.getEditProduct);
 
-// router.post('/edit-product', adminController.postEditProduct);
+router.post('/edit-product', adminController.postEditProduct);
 
-// router.post('/delete-product', adminController.postDeleteProduct);
+router.post('/delete-product', adminController.postDeleteProduct);
 
 module.exports = router;
